+++ libs/angular-accelerator/src/lib/components/search-header/search-header.component.ts
@@ -12,6 +12,11 @@ import {
 import { Action } from '../page-header/page-header.component'
 import { SearchConfigInfo } from '../../model/search-config-info'
 
+export interface SearchHeaderComponentState {
+  activeViewMode?: 'basic' | 'advanced'
+  selectedSearchConfig?: SearchConfigInfo
+}
+
 /**
  * To trigger the search when Enter key is pressed inside a search parameter field,
  * an EventListener for keyup enter event is added for HTML elements which have an input.
@@ -55,6 +60,7 @@ export class SearchHeaderComponent implements AfterViewInit {
   @Output() resetted: EventEmitter<any> = new EventEmitter()
   @Output() selectedSearchConfigChanged: EventEmitter<SearchConfigInfo> = new EventEmitter()
   @Output() viewModeChanged: EventEmitter<'basic' | 'advanced'> = new EventEmitter()
+  @Output() componentStateChanged: EventEmitter<SearchHeaderComponentState> = new EventEmitter()
   @ContentChild('additionalToolbarContent')
   additionalToolbarContent: TemplateRef<any> | undefined
 
@@ -80,6 +86,9 @@ export class SearchHeaderComponent implements AfterViewInit {
   toggleViewMode() {
     this.viewMode = this.viewMode === 'basic' ? 'advanced' : 'basic'
     this.viewModeChanged?.emit(this.viewMode)
+    this.componentStateChanged.emit({
+      activeViewMode: this.viewMode
+    })
     this.updateHeaderActions()
     setTimeout(() => this.addKeyUpEventListener())
   }
@@ -130,5 +139,8 @@ export class SearchHeaderComponent implements AfterViewInit {
 
   confirmSearchConfig(searchConfig: SearchConfigInfo) {
     this.selectedSearchConfigChanged?.emit(searchConfig)
+    this.componentStateChanged.emit({
+      selectedSearchConfig: searchConfig
+    })
   }
 }