+++ libs/ngrx-accelerator/src/lib/utils/effects/filter-for-query-params-changed.ts
@@ -15,9 +15,6 @@ export function filterForQueryParamsChanged<A extends RouterNavigatedAction>(
   return filterOutQueryParamsHaveNotChanged(router, queryParamsTypeDef, allowEmptyQueryParamsList)
 }
 
-/**
- * @deprecated Please import from \@onecx/ngrx-accelerator
- */
 export function filterOutQueryParamsHaveNotChanged<A extends RouterNavigatedAction>(
   router: Router,
   queryParamsTypeDef: ZodType,